renaming env.py to state.py
